ml/api/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
pillow
streamlit
requests
